+++ packages/expo-router/src/__tests__/getRoutes.test.node.ts
@@ -4,6 +4,7 @@ import {
   getUserDefinedDeepDynamicRoute,
   getRecursiveTree,
   FileNode,
+  assertDuplicateRoutes,
 } from "../getRoutes";
 import { RequireContext } from "../types";
 
@@ -63,6 +64,32 @@ const asRouteNode = (route: Partial<RouteNode>) => ({
   route: "INVALID_TEST_VALUE",
   contextKey: "INVALID_TEST_VALUE",
   ...route,
+});
+
+describe(assertDuplicateRoutes, () => {
+  const originalEnv = process.env.NODE_ENV;
+
+  afterEach(() => {
+    process.env.NODE_ENV = originalEnv;
+  });
+  it(`throws if there are duplicate routes`, () => {
+    expect(() =>
+      assertDuplicateRoutes(["a.js", "a.tsx", "b.js"])
+    ).toThrowErrorMatchingInlineSnapshot(
+      `"Multiple files match the route name \\"a\\"."`
+    );
+  });
+
+  it(`doesn't throw if there are no duplicate routes`, () => {
+    expect(() =>
+      assertDuplicateRoutes(["a", "b", "/c/d/e.js", "f/g.tsx"])
+    ).not.toThrow();
+  });
+
+  it(`doesn't throw if running in production`, () => {
+    process.env.NODE_ENV = "production";
+    expect(() => assertDuplicateRoutes(["a", "a.js"])).not.toThrow();
+  });
 });
 
 describe(getRecursiveTree, () => {
